refactor: 리포트 도메인 서버 빌더 분리
