+++ src/pages/Report/lib/toReportPageData.ts
@@ -1,21 +1,12 @@
-import {
-  pickBoolean,
-  pickSourceNumber,
-  pickSourceRecord,
-  pickSourceString,
-  pickSourceStringArray,
-} from '@/shared/api/responseAccess/payloadAccess';
-import type { ResolvedScanUrls } from '@/shared/lib/scan-session/scanUrlResolution';
+import { pickSourceNumber, pickSourceStringArray } from '@/shared/api/responseAccess/payloadAccess';
 import type { ScanSessionSnapshot } from '@/shared/store/scanSessionStore';
 import type { ResultTone } from '@/shared/types/resultTone';
 
+import { buildReportDomainComparison } from './reportDomainData';
 import { createReportPageDataContext } from './reportPageDataContext';
 import { buildReportReputation } from './reportReputationData';
-import {
-  missingReportInfoLabel,
-  reportFallbackCopyByTone,
-  trustScoreFallbackByTone,
-} from '../constants/reportText';
+import { buildReportServerInfo, resolveReportServerInfoRecords } from './reportServerInfoData';
+import { reportFallbackCopyByTone, trustScoreFallbackByTone } from '../constants/reportText';
 
 import type { ReportPageData } from '../types/reportPage.types';
 
@@ -48,51 +39,6 @@ function formatDateLabel(rawDate: string | null, fallbackLabel = '분석 시각 
   return `${year}-${month}-${day} ${hours}:${minutes}`;
 }
 
-function formatCertificateStatus(certificateRecord: Record<string, unknown> | null): string | null {
-  const isValid = pickBoolean(certificateRecord, ['valid', 'isValid', 'is_valid']);
-
-  if (isValid === null) {
-    return null;
-  }
-
-  return isValid ? '유효한 인증서' : '유효하지 않은 인증서';
-}
-
-function formatCertificateValidityPeriod(
-  certificateRecord: Record<string, unknown> | null,
-): string | null {
-  const validFrom = pickSourceString([certificateRecord], ['validFrom', 'valid_from']);
-  const validTo = pickSourceString([certificateRecord], ['validTo', 'valid_to']);
-
-  if (!validFrom && !validTo) {
-    return null;
-  }
-
-  return `${validFrom ?? '알 수 없음'} - ${validTo ?? '알 수 없음'}`;
-}
-
-function resolveCertificateTone(
-  rawStatusText: string | null,
-): ReportPageData['serverInfo']['certificateStatusTone'] {
-  if (!rawStatusText) {
-    return 'warning';
-  }
-
-  const normalizedStatusText = rawStatusText.trim().toLowerCase();
-  const errorSignals = ['expired', 'invalid', 'revoked', '만료', '폐기', '유효하지 않'];
-  const successSignals = ['valid', 'active', '유효', '정상'];
-
-  if (errorSignals.some((signal) => normalizedStatusText.includes(signal))) {
-    return 'error';
-  }
-
-  if (successSignals.some((signal) => normalizedStatusText.includes(signal))) {
-    return 'success';
-  }
-
-  return 'warning';
-}
-
 function resolveDetectedRiskTypes(sources: unknown[], riskLevel: ResultTone): string[] {
   const riskTypes = pickSourceStringArray(
     sources,
@@ -101,97 +47,6 @@ function resolveDetectedRiskTypes(sources: unknown[], riskLevel: ResultTone): st
   );
 
   return riskTypes.length > 0 ? riskTypes : reportFallbackCopyByTone[riskLevel].detectedRiskTypes;
-}
-
-function resolveUrlComparisonSummary({ destinationUrl, originalUrl }: ResolvedScanUrls): string {
-  if (originalUrl && destinationUrl && originalUrl !== destinationUrl) {
-    return '스캔된 QR URL이 최종 목적지와 다른 주소로 리다이렉트됩니다.';
-  }
-
-  return '스캔된 QR URL과 최종 목적지가 동일합니다.';
-}
-
-function resolveServerInfoRecord(rawServerInfoRecord: Record<string, unknown> | null) {
-  if (!rawServerInfoRecord) {
-    return {
-      certificateRecord: null,
-      serverInfoRecord: null,
-    };
-  }
-
-  const certificateRecord = pickSourceRecord([rawServerInfoRecord], ['certificate']);
-
-  return {
-    certificateRecord,
-    serverInfoRecord: {
-      ...rawServerInfoRecord,
-      certificateStatusText:
-        pickSourceString(
-          [rawServerInfoRecord],
-          ['certificateStatusText', 'certificate_status_text'],
-        ) ?? formatCertificateStatus(certificateRecord),
-      certificateValidityPeriod:
-        pickSourceString(
-          [rawServerInfoRecord],
-          ['certificateValidityPeriod', 'certificate_validity_period'],
-        ) ?? formatCertificateValidityPeriod(certificateRecord),
-      serverLocation: pickSourceString(
-        [rawServerInfoRecord],
-        ['serverLocation', 'server_location', 'location'],
-      ),
-      serverType: pickSourceString([rawServerInfoRecord], ['serverType', 'server_type', 'type']),
-    },
-  };
-}
-
-function buildDomainComparison(
-  domainComparisonRecord: Record<string, unknown> | null,
-  riskLevel: ResultTone,
-  urls: ResolvedScanUrls,
-): ReportPageData['domainComparison'] {
-  return {
-    officialUrl:
-      pickSourceString([domainComparisonRecord], ['officialUrl', 'official_url']) ??
-      urls.destinationUrl,
-    riskBadgeText:
-      pickSourceString([domainComparisonRecord], ['riskBadgeText', 'risk_badge_text']) ??
-      reportFallbackCopyByTone[riskLevel].domainRiskBadge,
-    summary:
-      pickSourceString([domainComparisonRecord], ['summary', 'description']) ??
-      resolveUrlComparisonSummary(urls),
-    suspiciousUrl:
-      pickSourceString([domainComparisonRecord], ['suspiciousUrl', 'suspicious_url']) ??
-      urls.originalUrl,
-  };
-}
-
-function buildServerInfo(
-  serverInfoRecord: Record<string, unknown> | null,
-  certificateRecord: Record<string, unknown> | null,
-): ReportPageData['serverInfo'] {
-  const certificateStatusText =
-    pickSourceString([serverInfoRecord], ['certificateStatusText', 'certificate_status_text']) ??
-    formatCertificateStatus(certificateRecord) ??
-    '확인 필요';
-
-  return {
-    certificateIssuer:
-      pickSourceString([serverInfoRecord], ['certificateIssuer', 'certificate_issuer']) ??
-      pickSourceString([certificateRecord], ['issuer']) ??
-      missingReportInfoLabel,
-    certificateStatusText,
-    certificateStatusTone: resolveCertificateTone(certificateStatusText),
-    certificateValidityPeriod:
-      pickSourceString(
-        [serverInfoRecord],
-        ['certificateValidityPeriod', 'certificate_validity_period'],
-      ) ?? missingReportInfoLabel,
-    serverLocation:
-      pickSourceString([serverInfoRecord], ['serverLocation', 'server_location']) ??
-      missingReportInfoLabel,
-    serverType:
-      pickSourceString([serverInfoRecord], ['serverType', 'server_type']) ?? missingReportInfoLabel,
-  };
 }
 
 export function toReportPageData(session: ScanSessionSnapshot): ReportPageData {
@@ -204,11 +59,12 @@ export function toReportPageData(session: ScanSessionSnapshot): ReportPageData {
     sources,
     urls,
   } = createReportPageDataContext(session);
-  const { certificateRecord, serverInfoRecord } = resolveServerInfoRecord(rawServerInfoRecord);
+  const { certificateRecord, serverInfoRecord } =
+    resolveReportServerInfoRecords(rawServerInfoRecord);
 
   return {
     detectedRiskTypes: resolveDetectedRiskTypes(sources, riskLevel),
-    domainComparison: buildDomainComparison(domainComparisonRecord, riskLevel, urls),
+    domainComparison: buildReportDomainComparison(domainComparisonRecord, riskLevel, urls),
     reputation: buildReportReputation(reputationRecord, internalDbRecord, sources),
     reportTitle: '상세 분석 리포트',
     riskDescription: reportFallbackCopyByTone[riskLevel].riskDescription,
@@ -216,7 +72,7 @@ export function toReportPageData(session: ScanSessionSnapshot): ReportPageData {
     riskLevelText: reportFallbackCopyByTone[riskLevel].riskLevelText,
     scannedAt: formatDateLabel(urls.scannedAt),
     scannedUrl: urls.scannedUrl,
-    serverInfo: buildServerInfo(serverInfoRecord, certificateRecord),
+    serverInfo: buildReportServerInfo(serverInfoRecord, certificateRecord),
     trustScore: clampTrustScore(
       pickSourceNumber(sources, ['trustScore', 'trust_score', 'score']),
       riskLevel,
